reinstate console click button
