+++ frontend/src/components/Gallery/AppBar/Platform/Base.vue
@@ -2,31 +2,45 @@
 import BaseGalleryAppBar from "@/components/Gallery/AppBar/Base.vue";
 import PlatformIcon from "@/components/common/Platform/Icon.vue";
 import FirmwareBtn from "@/components/Gallery/AppBar/Platform/FirmwareBtn.vue";
-import SettingsBtn from "@/components/Gallery/AppBar/Platform/SettingsBtn.vue";
 import FirmwareDrawer from "@/components/Gallery/AppBar/Platform/FirmwareDrawer.vue";
 import PlatformInfoDrawer from "@/components/Gallery/AppBar/Platform/PlatformInfoDrawer.vue";
 import storeRoms from "@/stores/roms";
+import storeNavigation from "@/stores/navigation";
 import { storeToRefs } from "pinia";
 
 // Props
 const romsStore = storeRoms();
+const navigationStore = storeNavigation();
 const { currentPlatform } = storeToRefs(romsStore);
+const { activePlatformInfoDrawer } = storeToRefs(navigationStore);
 </script>
+
 <template>
   <base-gallery-app-bar :show-playables-filter="false" show-filter-bar>
     <template #prepend>
-      <platform-icon
+      <v-btn
         v-if="currentPlatform"
-        :slug="currentPlatform.slug"
-        :name="currentPlatform.name"
-        :fs-slug="currentPlatform.fs_slug"
-        :size="36"
-        class="mx-3"
-      />
+        variant="text"
+        rounded="0"
+        icon="mdi-cog"
+        :color="activePlatformInfoDrawer ? 'primary' : ''"
+        @click="navigationStore.switchActivePlatformInfoDrawer()"
+      >
+        <platform-icon
+          :slug="currentPlatform.slug"
+          :name="currentPlatform.name"
+          :fs-slug="currentPlatform.fs_slug"
+          :size="36"
+          class="mx-3"
+        />
+        <v-icon
+          icon="mdi-cog"
+          size="xs"
+          class="position-absolute"
+          style="bottom: 4px; right: 4px"
+        ></v-icon>
+      </v-btn>
       <firmware-btn />
-    </template>
-    <template #append>
-      <settings-btn />
     </template>
   </base-gallery-app-bar>
